Add excel file used to build the graph 1
The number of armed violence incidents increased until August 2018, and then decreased until November 2018. It then remained quite constant until July 2019, except from a peak in March 2019. Finally, violence seems to be decreasing in the last few months. But for the end of 2019, the month of December is not complete yet so the value is not correct nor relevant. The decrease observed from July 2019 is also to be taken with caution because it could be explained by the time that it takes to the team of Fogo Cruzado to collect, verify, and publish the data.