Add qgis files used to build figure 3
Location of armed violence incidents in the city of Rio de Janeiro in 2018 and 2019

Armed violence remains concentrated in the most dense areas of the city, mainly in the Zona Norte, which concentrates precarious, segregated and isolated neighborhoods where drugs gangs are ruling the territory. There is much less violence in the Zona Oeste, where favelas are mostly ruled by extortion milicias rather than drug gangs. This has been indeed proven in many studies of criminal governance. We can also see that the touristic and economic center of the city, the Zona Sul, is not exempted from armed violence, but the incidents are mostly occurring in the hills where favelas are located.
Between 2018 and 2019, the only evolution concerns the Zona Oeste where there has been slightly less incidents this year. This again could be explained by the presence of militias in these neighborhoods, and their reinforced links with the new governor Witzel and his team.