Add csv file database Fogo Cruzado
To explore the evolution of urban violence in Rio de Janeiro, we used data produced by the Brazilian organization Fogo Cruzado created in 2015 and operating since 2017. It’s a collaborative digital platform registering all the incidences of armed violence in the metropolitan region of Rio de Janeiro through various sources : an app where any user can signal an event, an informants network of collectives and residents all over the territory, as well as press and authority informations. All these sources are crossed to obtained the more complete data about the occurrences of armed violence, the location, the date and the number of deaths and injured, either civilians or police.
Fogo Cruzado recently opened and API for open consultation of all the data they have. Indeed, they want to be ‘’transparent, collaborative and innovative’’. Nevertheless it is worth noting that it is necessary to login with an email address to access the data.
https://fogocruzado.org.br/sobre-a-api/
The data we could obtain did not have any issue and looked reliable so we did not have any work to clean it.